fix warnings
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@3684 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/ejb/services/ServiceTimerSessionBean.java
@@ -164,7 +164,7 @@ public class ServiceTimerSessionBean extends BaseSessionBean implements javax.ej
 		Integer timerInfo = (Integer) timer.getInfo();
 		
 		
-	    ServiceConfiguration serviceData = getServiceSession().getServiceConfiguration(intAdmin, timerInfo);
+	    ServiceConfiguration serviceData = getServiceSession().getServiceConfiguration(intAdmin, timerInfo.intValue());
 	    if(serviceData != null){
 	    	String serviceName = getServiceSession().getServiceName(intAdmin, timerInfo.intValue());
 	    	IWorker worker = getWorker(serviceData,serviceName);
